+++ fronted/src/components/ReceivedRequestsList.jsx
@@ -17,7 +17,8 @@ function ReceivedRequestsList({ handleShowToast }) {
         setUsers(usersResponse.data);
 
         const requestsResponse = await api.get('api/game/game-requests/');
-        const filteredRequests = requestsResponse.data.filter(request => request.requestee === user.id);
+        // Filtrar solicitudes por requestee, initiated false y accepted false
+        const filteredRequests = requestsResponse.data.filter(request => request.requestee === user.id && !request.initiated && !request.accepted);
 
         const requestsWithUsernames = filteredRequests.map(request => {
           const requester = usersResponse.data.find(u => u.id === request.requester);
@@ -37,12 +38,14 @@ function ReceivedRequestsList({ handleShowToast }) {
   }, [handleShowToast, user]);
 
   const handleDelete = async (id) => {
-    try {
-      await api.delete(`api/game/game-requests/${id}/`);
-      setRequests(requests.filter(request => request.id !== id));
-      handleShowToast("Request Rejected.");
-    } catch (error) {
-      setErrors({ delete: error.response.data.detail });
+    if (window.confirm("¿Está seguro de que desea eliminar esta solicitud? Esto también eliminará el juego asociado.")) {
+      try {
+        await api.delete(`api/game/game-requests/${id}/`);
+        setRequests(requests.filter(request => request.id !== id));
+        handleShowToast("Request Rejected.");
+      } catch (error) {
+        setErrors({ delete: error.response.data.detail });
+      }
     }
   };
 
@@ -53,24 +56,24 @@ function ReceivedRequestsList({ handleShowToast }) {
   return (
     <div className="d-flex justify-content-center">
       <div className="w-100" style={{ maxWidth: '600px' }}>
-        <h2>Received Requests</h2>
+        <h2>Solicitudes</h2>
         {errors.fetch && <p className="text-danger">{errors.fetch}</p>}
         {requests.length === 0 ? (
-          <p>No game requests yet.</p>
+          <p>Parece que nadie te ha invitado a jugar</p>
         ) : (
           <ul className="list-group">
             {requests.map(request => (
               <li className="list-group-item d-flex justify-content-between align-items-center" key={request.id}>
                 <div>
-                  <p><strong>Requester:</strong> {request.requesterUsername}</p>
-                  <p><strong>Date:</strong> {new Date(request.created_at).toLocaleString()}</p>
+                  <p><strong>Solicitante:</strong> {request.requesterUsername}</p>
+                  <p><strong>Fecha:</strong> {new Date(request.created_at).toLocaleString()}</p>
                 </div>
                 <div>
                   <button className="btn btn-success me-2" onClick={() => handleAcceptRedirect(request.id)}>
-                    <i className="bi bi-play"></i> Accept
+                    <i className="bi bi-play"></i> Aceptar
                   </button>
                   <button className="btn btn-danger" onClick={() => handleDelete(request.id)}>
-                    <i className="bi bi-trash3"></i> Reject
+                    <i className="bi bi-trash3"></i> Rechazar
                   </button>
                   {errors.delete && <p className="text-danger">{errors.delete}</p>}
                 </div>
